Refactor job service and provider
