+++ rundeckapp/grails-app/controllers/rundeck/controllers/MenuController.groovy
@@ -3131,7 +3131,7 @@ Since: v17''',
         }
 
 
-        def list = jobSchedulesService.getAllScheduled(uuid)
+        def list = jobSchedulesService.getAllScheduled(uuid, null)
         //filter authorized jobs
         Map<String, UserAndRolesAuthContext> projectAuths = [:]
         def authForProject = { String project ->
