+++ rundeckapp/src/main/groovy/org/rundeck/app/data/job/converters/OrchestratorUpdater.groovy
@@ -1,13 +1,13 @@
 package org.rundeck.app.data.job.converters
 
 import com.fasterxml.jackson.databind.ObjectMapper
+import org.rundeck.app.data.model.v1.job.orchestrator.OrchestratorData
 import rundeck.Orchestrator
-import rundeck.data.job.RdOrchestrator
 
 class OrchestratorUpdater {
     static ObjectMapper mapper = new ObjectMapper()
 
-    static void updateOrchestrator(Orchestrator orchestrator, RdOrchestrator rdo) {
+    static void updateOrchestrator(Orchestrator orchestrator, OrchestratorData rdo) {
         orchestrator.type = rdo.type
         orchestrator.content = mapper.writeValueAsString(rdo.configuration)
     }
